Make all values unknown by default
For sensors of class "total increasing", we need to avoid any bogus
readings which sould reset Home assistants idea of the last sensor
value. For example if using the cumulative "ac energy" for the energy
integration, it resetting to 1 and then back to it's true value will
cause erroneous steps in the energy monitor.
Import STATE_UNKNOWN and set all defaults values to it.