Setup additional device info (model, serial number, sw version)
Read and add additional device information to the hub using the
parameters read from the hub.

With this change, migrate the device to use the serial number rather
than the name of the device as the unique id.  To facilitate this, split
the current `SolaredgeModbusHub` into the `Hub` functionality, which
handles reading from the device, and the `SolaredgeModbusCoordinator`,
which handles the Home Assistant updates by reading from the `Hub`.
This allows the hub to be instatiated in the migration step, which is
needed to move the entities over to have the new unique id.

For some reason, doing this migration leaves the old device with no
entities in it, so add an empty `async_remove_config_entry_device`
function, which makes a "Delete" button show up on the device, allowing
the orphan device to be deleted from the UI.